Implement simple change calculator with bugs
I know what they are, but if I explain them here that'd defeat the
purpose of the exercise, sorry!  That said, in general it's a good idea to
check the commit messages just in case there are clues, so you get credit
for looking, even if in this case it's not useful.